[BACKEND] feature(user): alocar interface de user service para o arquivo de implementação
