+++ backend/src/core/auth/service/auth-impl.service.ts
@@ -3,9 +3,9 @@ import { FirebaseService } from 'src/firebase/firebase.service';
 import { AuthService } from './impl/auth.service';
 import { UserData } from 'src/core/user/user.entity';
 import { SignInDto } from '../dtos/sign-in.dto';
-import { UserService } from 'src/core/user/service/user.service';
 import { UserDto } from 'src/core/user/dtos/user-dto';
 import { SignInResponseDto } from '../dtos/sign-in-response-dto';
+import { UserService } from 'src/core/user/service/impl/user-impl.service';
 
 @Injectable()
 export class AuthServiceImpl implements AuthService {
