+++ backend/src/core/user/service/impl/user-impl.service.ts
@@ -1,7 +1,21 @@
 import { Injectable } from '@nestjs/common';
 import { UserRepository } from '../../user.repository';
 import { UserData, UserEntity } from '../../user.entity';
-import { UserService } from '../user.service';
+
+export abstract class UserService {
+  public abstract getUserById(id: number): Promise<UserEntity | null>;
+
+  public abstract getUserByEmail(email: string): Promise<UserEntity | null>;
+
+  public abstract findAll(): Promise<UserEntity[]>;
+
+  public abstract create(user: UserData): Promise<UserEntity>;
+
+  public abstract update(
+    id: number,
+    user: UserData,
+  ): Promise<UserEntity | null>;
+}
 
 @Injectable()
 export class UserServiceImpl implements UserService {
